- Fix NtSetInformationThread(ThreadBasePriority) - Fix buffer overflow in KeRosDumpStackFrames - Fix usage of garbage/uninitialized handle entry local variable in ObpCreateHandle - Fix buffer overflow and logic bug in IoConnectInterrupt - Fix MSVC warnings
svn path=/trunk/; revision=28153

+++ reactos/ntoskrnl/ke/i386/kiinit.c
@@ -216,10 +216,11 @@ KiInitMachineDependent(VOID)
                                               CurrentSample->TSCStart;
 
                     /* Compute CPU Speed */
-                    CurrentSample->MHz = ((CurrentSample->TSCDelta *
-                                           CurrentSample->PerfFreq.QuadPart +
-                                           500000) /
-                                          (CurrentSample->PerfDelta * 1000000));
+                    CurrentSample->MHz = (ULONG)((CurrentSample->TSCDelta *
+                                                  CurrentSample->
+                                                  PerfFreq.QuadPart + 500000) /
+                                                 (CurrentSample->PerfDelta *
+                                                  1000000));
 
                     /* Check if this isn't the first sample */
                     if (Sample)
